requirements.txt
absl-py==0.13.0
astunparse==1.6.3
atari-py==0.2.9
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cloudpickle==1.6.0
gast==0.3.3
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.38.0
gym==0.18.3
h5py==2.10.0
idna==2.10
importlib-metadata==4.5.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.15
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.16.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
